backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy>=2.0.0
pyodbc>=5.0.0
python-dotenv>=1.0.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.0.0
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.6
azure-identity>=1.15.0
openai>=1.0.0
xai-sdk>=0.1.0
azure-storage-blob>=12.0.0

# Test dependencies
pytest>=9.0.0
pytest-asyncio>=0.23.0
httpx>=0.27.0,<0.28
